Adding DeferredTensor and DeferredVariable classes.
To support eager mode, we'd like to replace many of the places that we use Tensors to instead use functions-returning-Tensors (so that we can differentiate through them). However, such functions can be a bit unwieldy. The DeferredTensor class tries to put a more user-friendly interface on functions returning Tensors, most significantly by implementing an "apply" function, and overloading some commonly-used operators.

The DeferredVariable class is a DeferredTensor that wraps a tf.Variable. The tf.Variable itself is owned by an EvaluationContext, and is created when the DeferredVariable.create() method is called (which should, once things are hooked up, happen inside RateMinimizationProblem).

At the moment, these classes are not used for anything. They'll be hooked up in a later CL.

PiperOrigin-RevId: 263001325