Adding extra checks that forbid ordinary Tensors from being passed as inputs to the rate helpers, in eager mode.
This will probably break most existing eager-mode code (graph-mode code will be unaffected), but will be *much* safer, since it will no longer be as easy for a user to pass in a constant Tensor, when they intended to pass in a variable.

In addition, it prevents __hash__ from being called on tf.Variables in eager mode inside the rate helpers. In the training code, constrained_optimizer.py is also changed to not create a set from a var_list, preventing __hash__ from being called on tf.Variables there, either.

Finally, updating Recall_constraint.ipynb to work with this change, and removing some whitespace from Wiki_toxicity_fairness.ipynb.

PiperOrigin-RevId: 293432154