Changing implementation of _RatioWeights so that it includes multiple columns.
Previously, for _RatioWeights included one set of weights (in a rank-1 Tensor), and BinaryClassificationTerm had two of them (one for positive predictions, and one for negative).

With this change, _RatioWeights now contains a rank-2 Tensor with multiple columns (two, for BinaryClassificationTerm). This is already slightly more efficient, but will make a bigger difference when support for multiclass is added, since we'll then have one _RatioWeights object with m columns, m being the number of classes, instead of m _RatioWeights objects.

PiperOrigin-RevId: 301425314