Refactoring Formulation class to inherit from tf.Module.
Inheriting from tf.Module makes variable tracking automatic. Since we don't need to keep track of variables anymore, state() has been changed to be a method instead of a property, which is understood to "evaluate" the state instead of just returning a Tensor (it still just returns a Tensor in the current implementations).

PiperOrigin-RevId: 342928615